Change withAddAccount to AddAccountScreen
#38
