+++ src/navigation/routeKeys.ts
@@ -6,6 +6,7 @@ export const permissionGrantingPromptScreen = 'PermissionsGrantingPrompt' as con
 export const dashboardNavigator = 'DashboardNavigator' as const;
 
 export const apiLoadingScreen = 'Api Loading' as const;
+export const addAccountScreen = 'AddAccountScreen' as const;
 
 export const dashboardScreen = 'Dashboard' as const;
 export const tipsScreen = 'Tips' as const;
