+++ src/screen/DashboardScreen.tsx
@@ -7,7 +7,6 @@ import {ChainApiContext} from 'context/ChainApiContext';
 import {BalanceContext} from 'context/BalanceContext';
 import {useAccounts} from 'context/AccountsContext';
 import FadeInAnimatedView from 'presentational/FadeInAnimatedView';
-import withAddAccount, {InjectedPropTypes as AddAccountInjectedPropTypes} from 'src/hoc/withAddAccount';
 import AccountTeaser from 'layout/AccountTeaser';
 import globalStyles from 'src/styles';
 import CouncilSummaryTeaser from 'layout/CouncilSummaryTeaser';
@@ -15,23 +14,23 @@ import TreasurySummaryTeaser from 'layout/TreasurySummaryTeaser';
 import TipsSummaryTeaser from 'layout/tips/TipsSummaryTeaser';
 import {CompositeNavigationProp, useNavigation} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
-import {councilScreen, tipsScreen, treasuryScreen} from 'src/navigation/routeKeys';
+import {addAccountScreen, councilScreen, tipsScreen, treasuryScreen} from 'src/navigation/routeKeys';
 import LoadingView from 'src/presentational/LoadingView';
 import NetworkSelect from 'src/layout/NetworkSelect';
 import SafeView, {noTopEdges} from 'presentational/SafeView';
 import {NetworkContext} from 'context/NetworkContext';
-import {DashboardStackParamList, DrawerParamList} from 'src/navigation/navigation';
+import {AppStackParamList, DashboardStackParamList, DrawerParamList} from 'src/navigation/navigation';
 
 type PropTypes = {
   navigation: CompositeNavigationProp<
-    StackNavigationProp<DashboardStackParamList>,
+    CompositeNavigationProp<StackNavigationProp<DashboardStackParamList>, StackNavigationProp<AppStackParamList>>,
     DrawerNavigationProp<DrawerParamList>
   >;
 };
 
 const AddIcon = (props: IconProps) => <Icon {...props} name="person-add-outline" />;
 
-function DashboardScreen({navigation, accountAddProps}: PropTypes & AddAccountInjectedPropTypes) {
+function DashboardScreen({navigation}: PropTypes) {
   const {accounts, isLoading} = useAccounts();
   const [networkSelectOpen, setNetworkSelectOpen] = useState(false);
 
@@ -55,7 +54,11 @@ function DashboardScreen({navigation, accountAddProps}: PropTypes & AddAccountIn
       <FadeInAnimatedView>
         {!account ? (
           <Layout style={styles.container} level="1">
-            <Button size="large" appearance="ghost" onPress={accountAddProps.open} accessoryLeft={AddIcon}>
+            <Button
+              size="large"
+              appearance="ghost"
+              onPress={() => navigation.navigate(addAccountScreen)}
+              accessoryLeft={AddIcon}>
               Add Account
             </Button>
           </Layout>
@@ -92,7 +95,7 @@ const styles = StyleSheet.create({
   divider: {height: 2},
 });
 
-export default withAddAccount(DashboardScreen);
+export default DashboardScreen;
 
 function DashboardHeaderRight() {
   const {show} = useContext(BalanceContext);
